feat: update color styles and improve component structure; refactor Logo and SidebarHeader components
