+++ src/app/globals.css
@@ -7,6 +7,7 @@
   --color-bg: #161A1D;
   --color-bgSecondary: #252836;
   --color-blue: #7CBAFD;
+  --color-yellow: #FFDF20;
 }
 
 :root {
